backend/requirements.txt
absl-py==2.4.0
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
dnspython==2.8.0
ecdsa==0.19.1
email_validator==2.1.1
fastapi==0.110.0
flatbuffers==25.12.19
fonttools==4.62.1
garminconnect==0.2.40
garth==0.6.3
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.0
greenlet==3.3.2
grpcio==1.78.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.27.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mediapipe==0.10.32
numpy==2.4.3
oauthlib==3.3.1
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
opencv-python-headless==4.9.0.80
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
proto-plus==1.27.1
protobuf==4.25.8
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.6.3
pydantic-settings==2.2.1
pydantic_core==2.16.3
pyparsing==3.3.2
pytest==8.0.2
pytest-asyncio==0.23.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.9
PyYAML==6.0.3
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.5
sqladmin==0.23.0
SQLAlchemy==2.0.48
starlette==0.36.3
tqdm==4.67.3
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.27.1
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
WTForms==3.1.2